[Draft] Remote storage reads based on oldest timestamp in local storage
Currently all read queries are simply pushed to remote storage.
We need instead to compare the oldest timestamp in local storage
with the query range lower boundary. If the oldest timestamp is
older than the from parameter, then there is no need for remote
storage calls.
This commit add FirstTime() in SeriesIterator interface,
and logic to use it.

Fixes #3041.

Signed-off-by: Thibault Chataigner <t.chataigner@criteo.com>